feat: logging and telemetry overhaul
Refactored logging to use `slog` for structured logging and integrated
with OpenTelemetry. Now the logging to standard output is human readable
using `lmittmann/tint`. Logs are multiplexed to standard output and/or
OTLP, depending on the configuration. Log verbosity can be selected via
`--log-level`. In our application code we use the `slog` logger now,
rather than the `logr` one. Not much in it, but this is slightly more
convenient to use.

The implementation of the Kubernetes controller has been moved to
a separate internal `tracker` package, out of the `main.go` file.

Two new metrics have been added to record each half of the process:

- `kube-manifests-exporter` runtime
- `kube-manifests` commit to flux apply time

in addition to our main metric which is the sum of those two.