feat(observability): fix disconnected tracing between monitoring tools
Previously, our observability stack had disconnected components - traces
couldn't be directly accessed from metrics or logs, requiring manual
correlation. Grafana supports such navigation - we just need to
configure it.

This change implements such linking. We configure the `tracesToLogsV2`
feature to search for logs with the right trace and span IDs, with a
slight time shift to account for the time it takes to process the trace.
We also set up the `exemplarTraceIdDestinations` in Mimir to allow for
direct navigation from metrics to traces.