feat(otel): use exponential histogram, add build info and standard otel resources
The fixed buckets of the explicit histograms we use currently make
monitoring the variable durations of our CI runs a bit tricky.
Fortunately, OpenTelemetry has another kind of histogram called an
[Exponential Histogram], which can offer better resolution and
dynamically adjusts to changing data (up to a configured maximum size),
such as if we improve/regress performance or if the data is different in
quiet periods.

These are represented in Prometheus/Mimir as [Native Histograms].

It can be bit confusing to understand - here's some further reading
[0] [1] [2].

We also add build info which we set in CI, and will be able to use
during our CD workflow, and some standard Otel resources/labels.

[Exponential Histogram]: https://opentelemetry.io/docs/specs/otel/metrics/data-model/
[Native Histograms]: https://prometheus.io/docs/specs/native_histograms/

[0]: https://dyladan.me/histograms/2023/05/04/exponential-histograms/
[1]: https://newrelic.com/blog/best-practices/opentelemetry-histograms
[2]: https://grafana.com/docs/mimir/latest/send/otel-exponential-histograms/