build: set version for pushes
We want to configure CD for `flux-commit-tracker`, so that we roll out
new versions to our internal deployments for every push to `main`. For
that, the builds need to report their version, so that our automation
can tell whether the versions it's rolling out are coming up OK - it
will add a selector to filter for the version which just got built.

We currently hardcode the version to `0.0.0`, which isn't going to work
for that purpose. Here we introduce a step to calculate a monotonic
version string from the current commit and ref.

We're still not creating a release, but we can use semver pre-releases
as a space for versions. The rough idea here is to generate versions
like `0.0.0-main-<run id>.<run attempt>.<sha>`. The ID and attempt are
what makes the version monotonic - the SHA is just there to make it
easier to go back to the change which triggered a build.

`git` ref names can contain character sequences that aren't valid as
semver pre-release identifiers, so we apply some sanitisation. This is
slight overkill as we'll only be using this for `main` pushes, but
it should avoid surprises if we ever generalise this.

All done using JS, as this was becoming a bit hairy to read or write in
a shell script.